requirements.txt
fastapi[standard]
python-decouple
pymongo[srv]
passlib[bcrypt]